Improve transpose multiply
While not stellar (we could avoid horizontal adds with enough care),
this does produce something that I think is on par with the SSE
autovectorizer.

I can't get this function to show up in the profile anymore, so either
it got small enough to become inlined or it's few enough cycles that it
no longer shows up.

There is maybe an argument for gemv with openblas or something (or
the accelerate framework if mac os is available, getting us AMX), but
if I can't see this function in a profile I'm willing to drop it for
now.